+++ modules/protocol/ts6-generic.c
@@ -591,7 +591,7 @@ m_mlock(struct sourceinfo *si, int parc, char *parv[])
 		return;
 	}
 
-	time_t ts = atol(parv[0]);
+	time_t ts = atoll(parv[0]);
 	if (ts > c->ts)
 		return;
 
@@ -619,7 +619,7 @@ static void
 m_tb(struct sourceinfo *si, int parc, char *parv[])
 {
 	struct channel *c = channel_find(parv[0]);
-	time_t ts = atol(parv[1]);
+	time_t ts = atoll(parv[1]);
 
 	if (c == NULL)
 		return;
@@ -650,8 +650,8 @@ m_etb(struct sourceinfo *si, int parc, char *parv[])
 			!(si->s->flags & SF_EOB) && c->topic != NULL)
 		return;
 
-	channelts = atol(parv[0]);
-	topicts = atol(parv[2]);
+	channelts = atoll(parv[0]);
+	topicts = atoll(parv[2]);
 	if (c->topic == NULL || channelts < c->ts || (channelts == c->ts && topicts > c->topicts))
 		handle_topic_from(si, c, parv[3], topicts, parv[parc - 1]);
 }
@@ -732,7 +732,7 @@ m_sjoin(struct sourceinfo *si, int parc, char *parv[])
 
 	// :origin SJOIN ts chan modestr [key or limits] :users
 	c = channel_find(parv[1]);
-	ts = atol(parv[0]);
+	ts = atoll(parv[0]);
 
 	if (!c)
 	{
@@ -835,7 +835,7 @@ m_join(struct sourceinfo *si, int parc, char *parv[])
 
 	// :user JOIN ts chan modestr [key or limits]
 	c = channel_find(parv[1]);
-	ts = atol(parv[0]);
+	ts = atoll(parv[0]);
 
 	if (!c)
 	{
@@ -950,7 +950,7 @@ m_nick(struct sourceinfo *si, int parc, char *parv[])
 
 		slog(LG_DEBUG, "m_nick(): new user on `%s': %s", s->name, parv[0]);
 
-		u = user_add(parv[0], parv[4], parv[5], NULL, NULL, NULL, parv[7], s, atoi(parv[2]));
+		u = user_add(parv[0], parv[4], parv[5], NULL, NULL, NULL, parv[7], s, atoll(parv[2]));
 		if (u == NULL)
 			return;
 
@@ -975,7 +975,7 @@ m_nick(struct sourceinfo *si, int parc, char *parv[])
 
 		slog(LG_DEBUG, "m_nick(): nickname change from `%s': %s", si->su->nick, parv[0]);
 
-		if (user_changenick(si->su, parv[0], atoi(parv[1])))
+		if (user_changenick(si->su, parv[0], atoll(parv[1])))
 			return;
 
 		/* It could happen that our PING arrived late and the
@@ -1006,7 +1006,7 @@ m_uid(struct sourceinfo *si, int parc, char *parv[])
 		s = si->s;
 		slog(LG_DEBUG, "m_uid(): new user on `%s': %s", s->name, parv[0]);
 
-		u = user_add(parv[0], parv[4], parv[5], NULL, parv[6], parv[7], parv[8], s, atoi(parv[2]));
+		u = user_add(parv[0], parv[4], parv[5], NULL, parv[6], parv[7], parv[8], s, atoll(parv[2]));
 		if (u == NULL)
 			return;
 
@@ -1050,7 +1050,7 @@ m_euid(struct sourceinfo *si, int parc, char *parv[])
 			parv[7],				// uid
 			parv[parc - 1],				// gecos
 			s,					// object parent (server)
-			atoi(parv[2]));				// hopcount
+			atoll(parv[2]));			// hopcount
 		if (u == NULL)
 			return;
 
@@ -1410,7 +1410,7 @@ m_signon(struct sourceinfo *si, int parc, char *parv[])
 		return;
 
 	// NICK
-	if (user_changenick(u, parv[0], atoi(parv[3])))
+	if (user_changenick(u, parv[0], atoll(parv[3])))
 		return;
 
 	handle_nickchange(u); // If they're logging out, this will bug them about identifying. Or something.
