+++ modules/protocol/unreal.c
@@ -764,7 +764,7 @@ m_mlock(struct sourceinfo *si, int parc, char *parv[])
 		return;
 	}
 
-	time_t ts = atol(parv[0]);
+	time_t ts = atoll(parv[0]);
 	if (ts > c->ts)
 		return;
 
@@ -922,7 +922,7 @@ m_sjoin(struct sourceinfo *si, int parc, char *parv[])
 	{
 		// :origin SJOIN ts chan modestr [key or limits] :users
 		c = channel_find(parv[1]);
-		ts = atol(parv[0]);
+		ts = atoll(parv[0]);
 
 		if (!c)
 		{
@@ -955,7 +955,7 @@ m_sjoin(struct sourceinfo *si, int parc, char *parv[])
 	{
 		// :origin SJOIN ts chan :users
 		c = channel_find(parv[1]);
-		ts = atol(parv[0]);
+		ts = atoll(parv[0]);
 
 		if (!c)
 		{
@@ -988,7 +988,7 @@ m_sjoin(struct sourceinfo *si, int parc, char *parv[])
 	else if (parc == 2)
 	{
 		c = channel_find(parv[1]);
-		ts = atol(parv[0]);
+		ts = atoll(parv[0]);
 		if (!c)
 		{
 			slog(LG_DEBUG, "m_sjoin(): new channel: %s (modes lost)", parv[1]);
@@ -1080,7 +1080,7 @@ m_uid(struct sourceinfo *si, int parc, char *parv[])
 				if (!inet_ntop(af, ipdata, ipstring, sizeof ipstring))
 					iplen = 0;
 		}
-		u = user_add(parv[0], parv[3], parv[4], vhost, iplen != 0 ? ipstring : NULL, parv[5], parv[parc - 1], s, atoi(parv[2]));
+		u = user_add(parv[0], parv[3], parv[4], vhost, iplen != 0 ? ipstring : NULL, parv[5], parv[parc - 1], s, atoll(parv[2]));
 		if (u == NULL)
 			return;
 
@@ -1165,7 +1165,7 @@ m_nick(struct sourceinfo *si, int parc, char *parv[])
 				if (!inet_ntop(af, ipdata, ipstring, sizeof ipstring))
 					iplen = 0;
 		}
-		u = user_add(parv[0], parv[3], parv[4], vhost, iplen != 0 ? ipstring : NULL, NULL, parv[parc - 1], s, atoi(parv[2]));
+		u = user_add(parv[0], parv[3], parv[4], vhost, iplen != 0 ? ipstring : NULL, NULL, parv[parc - 1], s, atoll(parv[2]));
 		if (u == NULL)
 			return;
 
@@ -1187,7 +1187,7 @@ m_nick(struct sourceinfo *si, int parc, char *parv[])
 			if (authservice_loaded && should_reg_umode(u))
 				sts(":%s SVS2MODE %s +r", nicksvs.nick, u->nick);
 		}
-		else if (u->ts > 100 && (time_t)atoi(parv[6]) == u->ts)
+		else if (u->ts > 100 && (time_t)atoll(parv[6]) == u->ts)
 			handle_burstlogin(u, NULL, 0);
 
 		handle_nickchange(u);
@@ -1205,7 +1205,7 @@ m_nick(struct sourceinfo *si, int parc, char *parv[])
 
 		realchange = irccasecmp(si->su->nick, parv[0]);
 
-		if (user_changenick(si->su, parv[0], atoi(parv[1])))
+		if (user_changenick(si->su, parv[0], atoll(parv[1])))
 			return;
 
 		// fix up +r if necessary -- jilles
